+++ extensions/last-commit-info/src/main/java/de/sovity/edc/extension/LastCommitInfoExtension.java
@@ -14,7 +14,7 @@
 
 package de.sovity.edc.extension;
 
-import org.eclipse.edc.connector.api.datamanagement.configuration.DataManagementApiConfiguration;
+import org.eclipse.edc.connector.api.management.configuration.ManagementApiConfiguration;
 import org.eclipse.edc.runtime.metamodel.annotation.Inject;
 import org.eclipse.edc.spi.system.ServiceExtension;
 import org.eclipse.edc.spi.system.ServiceExtensionContext;
@@ -24,7 +24,7 @@ public class LastCommitInfoExtension implements ServiceExtension {
 
     public static final String EXTENSION_NAME = "LastCommitInfoExtension";
     @Inject
-    private DataManagementApiConfiguration config;
+    private ManagementApiConfiguration config;
     @Inject
     private WebService webService;